Made a new widget for the columns in the weekplan screen, which is needed for testing and to make the code more readable
